sort items in EditPolicy.type_widgets before use
The keys in EditPolicy.type_widgets are keys that can inherit from each
other. Sort the items in the dict topologically (by class inheritance)
before using it for widget type resolution to ensure subclasses come
first. This allows for example to define a widget for
sqlalchemy.types.Text which will take precedence before the one for
sqlalchemy.types.String.